+++ src/DbEx.SqlServer/Console/SqlServerMigrationConsole.cs
@@ -48,15 +48,18 @@ public sealed class SqlServerMigrationConsole : MigrationConsoleBase<SqlServerMi
     /// Writes the supported <see cref="MigrationCommand.Script"/> help content.
     /// </summary>
     public void WriteScriptHelp()
-    { 
+    {
+        Logger?.LogInformation("{help}", "Inspect command and argument(s):");
+        Logger?.LogInformation("{help}", "  inspect <schema> <table1> <table2>...  Inspects one or more database tables within the specified schema.");
+        Logger?.LogInformation("{help}", string.Empty);
         Logger?.LogInformation("{help}", "Script command and argument(s):");
-        Logger?.LogInformation("{help}", "  script [default]                  Creates a default (empty) SQL script.");
-        Logger?.LogInformation("{help}", "  script alter <Schema> <Table>     Creates a SQL script to perform an ALTER TABLE.");
-        Logger?.LogInformation("{help}", "  script cdc <Schema> <Table>       Creates a SQL script to turn on CDC for the specified table.");
-        Logger?.LogInformation("{help}", "  script cdcdb                      Creates a SQL script to turn on CDC for the database.");
-        Logger?.LogInformation("{help}", "  script create <Schema> <Table>    Creates a SQL script to perform a CREATE TABLE.");
-        Logger?.LogInformation("{help}", "  script outbox <Schema> <Table>    Creates a SQL script to perform a CREATE TABLE(s) for an Outbox.");
-        Logger?.LogInformation("{help}", "  script refdata <Schema> <Table>   Creates a SQL script to perform a CREATE TABLE as reference data.");
-        Logger?.LogInformation("{help}", "  script schema <Schema>            Creates a SQL script to perform a CREATE SCHEMA.");
+        Logger?.LogInformation("{help}", "  script [default]                       Creates a default (empty) SQL script.");
+        Logger?.LogInformation("{help}", "  script alter <Schema> <Table>          Creates a SQL script to perform an ALTER TABLE.");
+        Logger?.LogInformation("{help}", "  script cdc <Schema> <Table>            Creates a SQL script to turn on CDC for the specified table.");
+        Logger?.LogInformation("{help}", "  script cdcdb                           Creates a SQL script to turn on CDC for the database.");
+        Logger?.LogInformation("{help}", "  script create <Schema> <Table>         Creates a SQL script to perform a CREATE TABLE.");
+        Logger?.LogInformation("{help}", "  script outbox <Schema> <Table>         Creates a SQL script to perform a CREATE TABLE(s) for an Outbox.");
+        Logger?.LogInformation("{help}", "  script refdata <Schema> <Table>        Creates a SQL script to perform a CREATE TABLE as reference data.");
+        Logger?.LogInformation("{help}", "  script schema <Schema>                 Creates a SQL script to perform a CREATE SCHEMA.");
     }
 }